Embark on the 4th DOM revision
This style will use raw pointers to nodes stored in arenas. Strings
will likewise be interned. It also creates a "thin" DOM to go along
with the typical "fat" DOM. In a fat DOM, every element has a
reference to the document from which it came. A thin DOM pushes that
responsibility onto the programmer, at the benefit of reduced size.